pyproject.toml
[tool.poetry]
name = "water-server-be"
version = "1.2.0"
description = ""
authors = ["shingaryu <shingaryu@gmail.com>"]
readme = "README.md"
#packages = [{include = "water_server_be"}]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "~2.2"
python-dotenv = "0.21.1"
waitress = "^2.1.2"
apscheduler = "~3.7.0"
line-bot-sdk = "^3.1.0"
google-api-python-client = "^2.102.0"
google-auth-oauthlib = "^1.1.0"
google-auth-httplib2 = "^0.1.1"

[tool.poetry.group.dev.dependencies]
pymongo = "^4.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
